refactor: provide internal X11 NET classes
Incorporate the NET parent and associated helper classes for internal
consumption instead of relying on KWindowSystem. We only need to include a
subset of funcitonality required for our window managing needs and few client
uses.

On the other side we continue to run the KWindowSystem plugin though in the
Wayland session for integration with Plasma.

+++ base/dbus/kwin.h
@@ -297,7 +297,7 @@ private:
             {QStringLiteral("caption"), win->meta.caption.normal},
             {QStringLiteral("clientMachine"), get_client_machine(*win)},
             {QStringLiteral("localhost"), is_local_host(*win)},
-            {QStringLiteral("type"), win->windowType()},
+            {QStringLiteral("type"), static_cast<int>(win->windowType())},
             {QStringLiteral("x"), win->geo.pos().x()},
             {QStringLiteral("y"), win->geo.pos().y()},
             {QStringLiteral("width"), win->geo.size().width()},
